auth5 mob/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
pyautogui
opencv-python
supabase
requests
runwayml
google-api-python-client
